fix: validation matches MTA-STS.
The previous version of this tool checks that the presented cert was
valid for at least one of the names specified in mx-hostnames.

The new validation rules ensure that:
    The selected MX host match at least one of the mx-hostnames, and
    The presented certificate is valid for the hostname.

We still need to talk about implementing API changes to forward all
relevant information to the front-end.

This API change also significantly bumps up the priority of getting
CI and unit-testing set up for this repo.